refactor(web): mood list 공용으로 변경
